collect processes and working directories

+++ ClangPowerTools/ClangPowerTools/Script/RunningProcesses.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 
 namespace ClangPowerTools
 {
@@ -8,17 +9,36 @@ namespace ClangPowerTools
     #region Members
 
     private List<Process> mProcesses = new List<Process>();
-
+    
     #endregion
 
     #region Public Methods
 
     public void Add(Process aProcess) => mProcesses.Add(aProcess);
 
+    public void KillById(int aId)
+    {
+      var procees = mProcesses.FirstOrDefault(p => p.Id == aId);
+      if (null == procees)
+        return;
+      procees.Kill();
+    }
+
     public void KillAll()
     {
       foreach (var process in mProcesses)
-        process.Kill();
+        if(!process.HasExited)
+          process.Kill();
+
+      List<Process> processes = new List<Process>();
+      processes.AddRange(Process.GetProcessesByName("clang++.exe"));
+      processes.AddRange(Process.GetProcessesByName("clang-tidy.exe"));
+
+      foreach (var process in processes)
+        if(!process.HasExited)
+          process.Kill();
+
+      mProcesses.Clear();
     }
 
     #endregion
